# github-repo/openapi-schema.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EvokeEssence Exchange API",
    "description": "API schema for EvokeEssence cryptocurrency exchange",
    "version": "1.0.0",
    "contact": {
      "name": "EvokeEssence s.r.o",
      "url": "https://evo-exchange.com",
      "email": "contact@evo-exchange.com"
    }
  },
  "servers": [
    {
      "url": "https://evo-exchange.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "summary": "User login",
        "description": "Authenticates a user and creates a session",
        "operationId": "login",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Username",
                    "example": "user123"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "User password",
                    "example": "SecurePassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "requiresTwoFactor": {
                      "type": "boolean",
                      "example": false
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "username": {
                          "type": "string",
                          "example": "user123"
                        },
                        "isAdmin": {
                          "type": "boolean",
                          "example": false
                        },
                        "isEmployee": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid username or password"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/logout": {
      "post": {
        "summary": "User logout",
        "description": "Logs out the current user and destroys the session",
        "operationId": "logout",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Logout successful"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user": {
      "get": {
        "summary": "Get current user",
        "description": "Returns details about the currently authenticated user",
        "operationId": "getCurrentUser",
        "tags": [
          "User"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 123
                    },
                    "username": {
                      "type": "string",
                      "example": "user123"
                    },
                    "email": {
                      "type": "string",
                      "format": "email",
                      "example": "user@example.com"
                    },
                    "fullName": {
                      "type": "string",
                      "example": "John Doe"
                    },
                    "isAdmin": {
                      "type": "boolean",
                      "example": false
                    },
                    "isEmployee": {
                      "type": "boolean",
                      "example": false
                    },
                    "kycStatus": {
                      "type": "string",
                      "example": "verified"
                    },
                    "balance": {
                      "type": "number",
                      "example": 10000.5
                    },
                    "balanceCurrency": {
                      "type": "string",
                      "example": "USD"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authenticated": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "User not authenticated"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Express session cookie for authentication"
      }
    },
...
  },
...
}